Concurrency Bugs
Discovered source of nasty race condition that was causing crash when
thread pools were used. Expression is NOT thread safe because variables
and values are stored. Instead, I will need to refactor the expression
api so that these values are passed down the evaluation chain so that
they are stored on each threads stack.