Working Unloading
Unloading works but there are some small performance issues. I intend to
rewrite the render stage so that it modifies the existing mesh if
possible instead of swapping in a new one. If there is not a noticeable
speed increase here, I intend to cache neighbors as they are computed in
order to speed up update stage.