Bug Fix and Performance
Finally solved bug that was causing square to get deleted and created
again in the same update pass. The first problem was that neighbor
squares weren't properly being checked for readiness in the
enableVertexNeighbor method. Secondly, when the data structure lacks
enough information to enable a vertex, the vertex is still conditionally
enabled and its square should not be merged if it is dependent on this
vertex. Unfortunately, performance issues still persist, and I have
determined the update stage is now the bottle neck. I suspect cacheing
neighbors will be a huge optimization so the plan is to proceed with
this next.