Optimization
First change here, neighbor cacheing is now fully functional and
properly removes squares, which gives it an edge over the previous tree
traversal methods. This has definitively resulted in a slight speed
increase for the update stage.

Regarding the render stage, it is very apparent that when working with
the JavaFx scene graph, modification should happen at the most primitive
level. I was previously swapping in a new triangle mesh for each updated
square's MeshView object. A quick optimization here involved retrieving
the existing mesh from the MeshView object and clearing all of its
values before rewriting them. This leaves less data behind for the
garbage collector to take care of and, what I believe is most important,
a new change listener need not be set up for every modified square.

I intend to start working on normal calculation soon, but it is
important to note that at this point, with the render stage as fast as
possible, the true bottleneck is the update stage. At a detail threshold
of 200, which is notably high for the noisy test function I am using,
update drags the frame rate down to 30 and bellow at close proximity.
This indicates that the update will need to occur on a separate thread.
For this to happen, all scene graph manipulation must be moved out of
the update stage and into the render stage, which will happen on the
application thread. This only really applies to mesh removal.