Changes To Render
Everything that modifies the JavaFx scene graph is now executed strictly
in the render stage. Unfortunately, I was using the JavaFx concurrency
package for creating new quadsquares. Now that I am moving the update
stage to its own thread, these concurrency objects throw exceptions
causing the update to fail. This is because JavaFx concurrency objects
are made to be modified strictly from the application thread. Next step
is switching all update related multithreaded tasks to the callable
interface instead of the JavaFx task. This will allow me to store
Futures in the quadSquareBuffer instead of references to the full task,
so the implementation will be a lot cleaner.