More Vertex Removal
Vertices that are part of a square that was generated but not connected
will now be removed if they aren't dependent. Also, if a square is
interrupted while being created, it will back track and delete vertices
that aren't dependent. All this is made possible through a new neighbor
method that allows the system to check these dependencies, even if the
square isn't actually part of the tree. The enableVertexNeighbor method
has been updated to make use of this as well.