Seams
Seams have been almost entirely eliminated but there are still a few
cases where this can occur, particularly when enabling/disabling a child
and recomputing its normals. One big optimization I would like to go
ahead and move forward with is directly computing triangle normals when
needed instead of storing them. I had originally done this to save some
computation time on the update stage but now that the update stage is
concurrent, memory takes precedence. This is especially true because the
next phase of this project will be texture generation, which will
unavoidably consume lots of memory.