Texture Work
Work on textures. Shading is baked using bilinear interpolation of
normals instead of strictly shade values being linearly interpolated
across triangles in the default JavaFx fragment shader. This
significantly reduces shading artifacts at the expense of increased time
in the update stage, but there are still some seams that need to be
eliminated.

At this point, factors in the internal API are causing both significant
memory and CPU overhead which produces periodic stalls. I plan to finish
the project using JavaFx as the primary rendering API and then moving
the algorithm to a more robust OpenGL engine that could realistically be
used for commercial applications.